+++ packages/file/lib/src/backends/record_replay/recording_random_access_file.dart
@@ -69,7 +69,7 @@ class RecordingRandomAccessFile extends Object
   String get identifier => '$runtimeType@$uid';
 
   @override
-  MutableRecording get recording => fileSystem.recording;
+  MutableRecording get recording => fileSystem.recording as MutableRecording;
 
   @override
   Stopwatch get stopwatch => fileSystem.stopwatch;